Fixes filename and folder problem
Previously files were just getting saves in the root directory of
our bucket. Solved this by adding today's date as a folder name an
prepending our filename with that. The next thing we did via
crypto.randomBytes is create a random string to save our file as on
AWS. The last thing we did was promisify both s3.upload and
crypto.randomBytes so both implement promises that are resolved or
rejected based on their success.